21. 8. 1. Sun) 01:25 mysite5_board 보안 처리

+++ src/main/java/com/javaex/controller/BoardController.java
@@ -23,52 +23,125 @@ public class BoardController {
 	
 	@RequestMapping(value="/list", method = { RequestMethod.GET, RequestMethod.POST })
 	public String list(Model model, @RequestParam(value = "keyword", required = false, defaultValue = "") String keyword) {
-		model.addAttribute("boardMap", boardService.boardMap(keyword));
+		model.addAttribute("boardMap", boardService.boardMap(keyword)); // --> boardDao.listCount(keyword), boardDao.boardList(keyword)
 		
 		return "/board/list";
 	}
 	
 	@RequestMapping(value="/writeForm", method = { RequestMethod.GET, RequestMethod.POST })
-	public String writeForm() {
+	public String writeForm(HttpSession session) {
+		// 로그인 정보가 없을때
 		
+		// Controller에서 처리
+		/*
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		
+		if (authUser != null) {
+			return "/board/writeForm";
+		} else {
+			return "redirect:/user/loginForm";
+		}
+		*/
+		
+		// wirteForm에서 처리
 		return "/board/writeForm";
+		
 	}
 	
 	@RequestMapping(value="/write", method = { RequestMethod.GET, RequestMethod.POST })
 	public String write(@ModelAttribute BoardVo boardVo, HttpSession session) {
-		boardVo.setUserNo(((UserVo)session.getAttribute("authUser")).getNo());
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
 		
-		boardService.write(boardVo);
+		if (authUser == null) { // 로그인 정보가 없을때
+			return "redirect:/user/loginForm";
+		}
+		
+		int authUserNo = authUser.getNo();
+		boardVo.setUserNo(authUserNo);
+		
+		boardService.write(boardVo); // boardVo = userNo, title, content
 		
 		return "redirect:/board/list";
 	}
 	
 	@RequestMapping(value="/read", method = { RequestMethod.GET, RequestMethod.POST })
-	public String read(Model model, @RequestParam("no") int no) {
-		model.addAttribute("getBoard", boardService.getBoardRead(no));
-
-		return "/board/read";
+	public String read(Model model, @RequestParam(value = "no", required = false, defaultValue = "0") int no) {
+		if (no == 0) { // read에 no파라미터가 입력되지 않았을때 0으로 처리하고 redirect:/board/list
+			return "redirect:/board/list";
+			
+		} else {
+			model.addAttribute("getBoard", boardService.getBoardRead(no));
+			
+			return "/board/read";
+		}
 	}
 	
 	@RequestMapping(value="/modifyForm", method = { RequestMethod.GET, RequestMethod.POST })
-	public String modifyForm(Model model, @RequestParam("no") int no) {
-		model.addAttribute("getBoard", boardService.getBoardModifyForm(no));
-
-		return "/board/modifyForm";
+	public String modifyForm(Model model, @RequestParam(value = "no", required = false, defaultValue = "0") int no, HttpSession session) {
+		
+		if (no == 0) { // modifyForm에 no파라미터가 입력되지 않았을때 0으로 처리하고 redirect:/board/list
+			return "redirect:/board/list";
+			
+		} else {
+			BoardVo boardVo = boardService.getBoardModifyForm(no);
+			int userNo = boardVo.getUserNo();
+			
+			UserVo authUser = (UserVo)session.getAttribute("authUser");		
+			
+			if (authUser == null) { // 로그인 정보가 없을때
+				return "redirect:/user/loginForm";
+			}
+			
+			int loginUserNo = authUser.getNo();
+			
+			if (userNo == loginUserNo) { // 수정하려는 글 작성자의 번호와 로그인 유저의 번호가 같을때
+				model.addAttribute("getBoard", boardVo);
+				
+				return "/board/modifyForm";
+				
+			} else { // 다른 유저의 글을 수정하려고 시도 할 때
+				return "redirect:/board/read?no=" + no;
+			}
+		}
+		
 	}
 	
 	@RequestMapping(value="/modify", method = { RequestMethod.GET, RequestMethod.POST })
 	public String modify(@ModelAttribute BoardVo boardVo) {
+		String boardVoTitle = boardVo.getTitle();
+		
+		if (boardVoTitle == null) {
+			return "redirect:/board/list";
+		}
+		
+		int no = boardVo.getNo();
+		
 		boardService.update(boardVo);
 		
-		return "redirect:/board/list";
+		return "redirect:/board/read?no=" + no;
 	}
 	
 	@RequestMapping(value="/delete", method = { RequestMethod.GET, RequestMethod.POST })
-	public String delete(@RequestParam("no") int no) {
-		boardService.delete(no);
+	public String delete(@RequestParam(value = "no", required = false, defaultValue = "0") int no, HttpSession session) {
 		
-		return "redirect:/board/list";
+		if (no == 0) {
+			return "redirect:/board/list";
+		}
+		
+		UserVo authUser = (UserVo)session.getAttribute("authUser");
+		int authUserNo = authUser.getNo();
+		
+		BoardVo getBoard = boardService.delete(no);
+		int boardUserNo = getBoard.getUserNo();
+		
+		if (authUserNo == boardUserNo) {
+			boardService.delete(no);
+			
+			return "redirect:/board/list";
+		} else {
+		
+			return "redirect:/board/list";
+		}
 	}
 
 }
